[Fix] MIDI mapping: Pattern record checkbox wasn't updated properly. [Fix] Sample map: was broken when dealing with sample indexes greater than 255. [Mod] Version: Changed version number. [Ref] Minor code cleanup.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@380 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/dlg_misc.cpp
@@ -2787,7 +2787,7 @@ LRESULT CSampleMapDlg::OnKeyboardNotify(WPARAM wParam, LPARAM lParam)
 				KeyboardMap[iNote] = pIns->Keyboard[iNote];
 			} else
 			{
-				KeyboardMap[iNote] = (BYTE)nSample;
+				KeyboardMap[iNote] = (WORD)nSample;
 			}
 /* rewbs.note: I don't think we need this with cust keys.
 // -> CODE#0009
